[unrated] Title: 문자열 정수의 합, Time: 0.03 ms, Memory: 76.9 MB -BaekjoonHub

+++ 프로그래머스/unrated/181849. 문자열 정수의 합/문자열 정수의 합.java
@@ -5,7 +5,7 @@ class Solution {
             
         for(int i = 0; i < num_str.length(); i++){
          a[i] = num_str.charAt(i); 
-           answer += ((int)a[i] - 48) % 10;
+           answer += (a[i] - 48) % 10;
         }  
         
         return answer;
